+++ src/main/java/com/suma/controller/AdvInfoContoller.java
@@ -1,11 +1,12 @@
 package com.suma.controller;
 
+import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.JSONObject;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-import com.suma.constants.CommonConstants;
 import com.suma.constants.ExceptionConstants;
 import com.suma.exception.AdvInfoException;
-import com.suma.exception.DefaultExceptionHandler;
+import com.suma.exception.AdvMaterialException;
 import com.suma.pojo.*;
 import com.suma.service.AdvInfoService;
 import com.suma.service.AdvMaterialService;
@@ -13,17 +14,18 @@ import com.suma.service.AdvTypeService;
 import com.suma.service.InfoMaterialService;
 import com.suma.utils.Result;
 import com.suma.vo.AdvInfoInsertVO;
-import com.suma.vo.AdvInfoQueryVO;
 import com.suma.vo.AdvInfoUpdateVO;
-import org.apache.commons.collections.map.HashedMap;
-import org.apache.commons.lang3.StringUtils;
+import com.suma.vo.AdvMaterialVO;
+import com.suma.vo.InfoMaterialVO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 
+import javax.sound.sampled.Line;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -31,14 +33,15 @@ import java.util.*;
 
 /**
  * 广告信息 增删改查
+ *
  * @auther: luxinzong
  * @date: 2018/10/15
  */
 @RestController
-@RequestMapping(value = "info",produces = "application/json;charset=utf-8")
-public class AdvInfoContoller extends BaseController{
+@RequestMapping(value = "info", produces = "application/json;charset=utf-8")
+public class AdvInfoContoller extends BaseController {
 
-    Logger logger = LoggerFactory.getLogger(AdvInfoContoller.class);
+    private static Logger logger = LoggerFactory.getLogger(AdvInfoContoller.class);
     @Autowired
     private AdvInfoService advInfoService;
 
@@ -51,36 +54,60 @@ public class AdvInfoContoller extends BaseController{
     @Autowired
     private AdvMaterialService advMaterialService;
 
+
+    /**
+     * 按照ID查找广告信息
+     * @param id
+     * @return
+     */
+    @RequestMapping(value = "queryById")
+    public Result query(Long id) {
+        return Result.success(advInfoService.findById(id));
+    }
+
     /**
      * 查询广告信息
+     *
      * @param
      * @return
      */
     @RequestMapping(value = "query", method = RequestMethod.GET)
-    public Result queryAdvInfo(Integer status,String name,String start,String end,
-                               Integer pageNum,Integer pageSize,String advType) throws ParseException {
-        if (pageNum == null || pageSize == null || status == null || advType == null) {//检查页码参数和广告状态是否为空
+    public Result queryAdvInfo(Integer status, String name, String start, String end,
+                               Integer pageNum, Integer pageSize, String advType) {
+        //检查是否缺少必须参数
+        if (pageNum == null || pageSize == null || status == null || advType == null) {
             throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_QUERYPARAMS_IS_NULL);
         }
+        //根据广告类型查询广告类型ID
         AdvTypeExample example1 = new AdvTypeExample();
         example1.createCriteria().andAdvtypeEqualTo(advType);
         List<AdvType> advTypes = advTypeService.selectByExample(example1);
+        //进行分页查询
         PageHelper.startPage(pageNum, pageSize);
+        //获取广告类型ID并存入集合中
         List<Long> list = new ArrayList<>();
         for (AdvType advType1 : advTypes) {
             System.out.println(advType1.getId());
             list.add(advType1.getId());
         }
+        //将查询参数存入map中
         Map<String, Object> map = new HashMap<>();
         map.put("name", name);
         map.put("status", status);
         map.put("startDate", start);
         map.put("endDate", end);
-        map.put("ids", list);
-        List<AdvInfo> advInfoList = advInfoService.selectAdvInfo(map);
         PageInfo<AdvInfo> listPageInfo = new PageInfo<>();
-        listPageInfo.setList(advInfoList);
-        listPageInfo.setTotal(advInfoList.size());
+        if (!CollectionUtils.isEmpty(list)) {
+            map.put("ids", list);
+            //根据map查询出广告信息
+            List<AdvInfo> advInfoList = advInfoService.selectAdvInfo(map);
+            //将广告信息和查询出来的总数存入PageInfo中返回给前端
+            listPageInfo.setList(advInfoList);
+            listPageInfo.setTotal(advInfoList.size());
+        } else {
+            List<AdvInfo> advInfoList = new ArrayList<AdvInfo>();
+            listPageInfo.setList(advInfoList);
+        }
         return Result.success(listPageInfo);
     }
 
@@ -91,24 +118,28 @@ public class AdvInfoContoller extends BaseController{
      * @return
      */
     @Transactional(rollbackFor = Exception.class)
-    @RequestMapping(value = "deleteAdvInfo", method = RequestMethod.POST)
+    @RequestMapping(value = "delete", method = RequestMethod.POST)
     public Result deleteAdvInfo(Long id) {
+        System.out.println(id);
         int a = advInfoService.deleteByPK(id);
         InfoMaterialExample example = new InfoMaterialExample();
         example.createCriteria().andAdvInfoIdEqualTo(id);
-        int b = infoMaterialService.deleteByExample(example);
-        return toResult(a*b);
+        if ( infoMaterialService.selectByExample(example).size() != 0) {
+            infoMaterialService.deleteByExample(example);
+        }
+        return Result.success();
     }
 
     /**
      * 批量删除广告信息
+     *
      * @param idsStr
      * @return
      */
     @Transactional(rollbackFor = Exception.class)
-    @RequestMapping(value = "delete", method = RequestMethod.POST)
+    @RequestMapping(value = "deleteAll", method = RequestMethod.POST)
     public Result deleteAdvInfos(String idsStr) {
-        String[] ids = idsStr.substring(0,idsStr.lastIndexOf(",")).split(",");
+        String[] ids = idsStr.substring(0, idsStr.lastIndexOf(",")).split(",");
         int count = 0;
         try {
             if (ids.equals("")) {
@@ -119,10 +150,12 @@ public class AdvInfoContoller extends BaseController{
                 advInfoService.deleteByPK(Long.valueOf(id));
                 InfoMaterialExample example = new InfoMaterialExample();
                 example.createCriteria().andAdvInfoIdEqualTo(Long.valueOf(id));
-                infoMaterialService.deleteByExample(example);
+                if ( infoMaterialService.selectByExample(example).size() != 0) {
+                    infoMaterialService.deleteByExample(example);
+                }
                 count++;
             }
-        }catch(Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
         if (count == ids.length) {
@@ -133,17 +166,19 @@ public class AdvInfoContoller extends BaseController{
     }
 
 
+
+
     /**
      * 创建广告信息
-     * @param advInfoInsertVO
+     * @param
      * @return
      */
     @Transactional(rollbackFor = Exception.class)
-    @RequestMapping(value = "select", method = RequestMethod.POST)
-    public Result insertAdvInfo (AdvInfoInsertVO advInfoInsertVO) throws ParseException {
-        System.out.println(advInfoInsertVO);
+    @RequestMapping(value = "insert")
+    public Result insertAdvInfo(String data) throws ParseException {
+        AdvInfoInsertVO advInfoInsertVO = JSON.parseObject(data, AdvInfoInsertVO.class);
         DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        DateFormat df = new SimpleDateFormat("HH:mm:ss");
+        DateFormat df = new SimpleDateFormat("HH:mm");
         AdvInfo advInfo = new AdvInfo();
         advInfo.setStartDate(dateFormat.parse(advInfoInsertVO.getStart()));
         advInfo.setEndDate(dateFormat.parse(advInfoInsertVO.getEnd()));
@@ -154,7 +189,7 @@ public class AdvInfoContoller extends BaseController{
         if (advInfo.getName() == null || advInfo.getStartDate() == null ||
                 advInfo.getEndDate() == null || advInfo.getPeriodTimeEnd() == null
                 || advInfo.getPeriodTimeStart() == null || advInfo.getStatus() == null
-                || advInfo.getMaterialType() == null) {
+                || advInfo.getMaterialType() == null || advInfo.getAdvTypeId() == null) {
             throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_MISSING_REQUIRED_PARAMS);
         }
         String name = advInfo.getName();
@@ -173,40 +208,59 @@ public class AdvInfoContoller extends BaseController{
         //获取广告信息ID
         Long advInfoId = advInfos.get(0).getId();
         //获取广告信息与资源关系表对象
-        InfoMaterial infoMaterial = new InfoMaterial();
+        List<InfoMaterialVO> infoMaterialVOS = advInfoInsertVO.getInfoMaterialVOS();
+
         //根据前端传递过来的文件名查找对应资源ID
-        AdvMaterialExample advMaterialExample = new AdvMaterialExample();
-        advMaterialExample.createCriteria().andFileNameEqualTo(advInfoInsertVO.getFileName());
-        List<AdvMaterial> advMaterials = advMaterialService.selectByExample(advMaterialExample);
-        Long materialId =  advMaterials.get(0).getId();
-        infoMaterial.setAdvInfoId(advInfoId);
-        infoMaterial.setMaterialId(materialId);
-        infoMaterial.setDuration(advInfoInsertVO.getDuration());
-        infoMaterial.setSequence(advInfoInsertVO.getSequence());
-        int b = infoMaterialService.save(infoMaterial);
-        return toResult(a*b);
+        AdvMaterialExample example1 = new AdvMaterialExample();
+        int count = 0;
+        for (InfoMaterialVO infoMaterialVO : infoMaterialVOS) {
+            example1.createCriteria().andFileNameEqualTo(infoMaterialVO.getFileName());
+            List<AdvMaterial> advMaterials = advMaterialService.selectByExample(example1);
+            if (advMaterials == null) {
+                throw new AdvMaterialException(ExceptionConstants.ADV_MATERIAL_IS_NULL);
+            }
+            Long advMaterialId = advMaterials.get(0).getId();
+            InfoMaterial infoMaterial = new InfoMaterial();
+            infoMaterial.setMaterialId(advMaterialId);
+            infoMaterial.setAdvInfoId(advInfoId);
+            BeanUtils.copyProperties(infoMaterialVO, infoMaterial);
+            int b = infoMaterialService.save(infoMaterial);
+            if (b > 0) {
+                count++;
+            }
+        }
+        if (count == infoMaterialVOS.size() && a > 0) {
+            return Result.success();
+        } else {
+            throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_INSERT_FAIL);
+        }
     }
 
     /**
      * 编辑广告信息
-     * @param advInfoUpdateVO
+     *
+     * @param data
      * @return
      */
     @Transactional(rollbackFor = Exception.class)
     @RequestMapping(value = "update", method = RequestMethod.POST)
-    public Result updateAdvInfo (AdvInfoUpdateVO advInfoUpdateVO) throws ParseException {
+    public Result updateAdvInfo(String data) throws ParseException {
+        AdvInfoUpdateVO advInfoUpdateVO = JSON.parseObject(data, AdvInfoUpdateVO.class);
+        System.out.println(advInfoUpdateVO);
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
         SimpleDateFormat sdf1 = new SimpleDateFormat("HH:mm:ss");
         AdvInfo advInfo = new AdvInfo();
         BeanUtils.copyProperties(advInfoUpdateVO, advInfo);
+        advInfo.setId(advInfoUpdateVO.getAdvInfoId());
         advInfo.setStartDate(sdf.parse(advInfoUpdateVO.getStart()));
         advInfo.setEndDate(sdf.parse(advInfoUpdateVO.getEnd()));
         advInfo.setPeriodTimeStart(sdf1.parse(advInfoUpdateVO.getPeriodTimeStart()));
         advInfo.setPeriodTimeEnd(sdf1.parse(advInfoUpdateVO.getPeriodTimeEnd()));
+        System.out.println(advInfo);
         if (advInfo.getId() == null || advInfo.getName() == null || advInfo.getStartDate() == null ||
                 advInfo.getEndDate() == null || advInfo.getPeriodTimeEnd() == null
                 || advInfo.getPeriodTimeStart() == null || advInfo.getStatus() == null
-                || advInfo.getMaterialType() == null) {
+                || advInfo.getMaterialType() == null || advInfo.getAdvTypeId() == null) {
             throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_MISSING_REQUIRED_PARAMS);
         }
         AdvInfoExample example = new AdvInfoExample();
@@ -215,21 +269,35 @@ public class AdvInfoContoller extends BaseController{
             throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_INFO_IS_NOT_EXIT);
         }
         //更新广告信息
-        int a =  advInfoService.update(advInfo);
-
+        int a = advInfoService.update(advInfo);
         //更新广告对应资源表信息
-        InfoMaterial infoMaterial = new InfoMaterial();
-        BeanUtils.copyProperties(advInfoUpdateVO, infoMaterial);
-
-        //根据文件名查找资源ID
+        //将前端传过来的数据复制到pojo中 advInfoUpdateVO -> infoMaterial中
         AdvMaterialExample example1 = new AdvMaterialExample();
-        example1.createCriteria().andFileNameEqualTo(advInfoUpdateVO.getFileName());
-        Long materialId = advMaterialService.selectByExample(example1).get(0).getId();
-        infoMaterial.setMaterialId(materialId);
-        InfoMaterialExample example2 = new InfoMaterialExample();
-        example2.createCriteria().andMaterialIdEqualTo(materialId).andAdvInfoIdEqualTo(advInfoUpdateVO.getAdvInfoId());
-        infoMaterialService.updateByExample(infoMaterial, example2);
-        return toResult(1);
+        List<InfoMaterialVO> infoMaterialVOS = advInfoUpdateVO.getInfoMaterialVOS();
+        int count = 0;
+        for (InfoMaterialVO infoMaterialVO : infoMaterialVOS) {
+            example1.createCriteria().andFileNameEqualTo(infoMaterialVO.getFileName());
+            List<AdvMaterial> advMaterials = advMaterialService.selectByExample(example1);
+            if (advMaterials == null) {
+                throw new AdvMaterialException(ExceptionConstants.ADV_MATERIAL_IS_NULL);
+            }
+            Long advMaterialId = advMaterials.get(0).getId();
+            InfoMaterial infoMaterial = new InfoMaterial();
+            infoMaterial.setMaterialId(advMaterialId);
+            infoMaterial.setAdvInfoId(advInfo.getId());
+            BeanUtils.copyProperties(infoMaterialVO, infoMaterial);
+            InfoMaterialExample example2 = new InfoMaterialExample();
+            example2.createCriteria().andMaterialIdEqualTo(advMaterialId).andAdvInfoIdEqualTo(advInfoUpdateVO.getAdvInfoId());
+            int b = infoMaterialService.updateByExample(infoMaterial, example2);
+            if (b > 0) {
+                count++;
+            }
+        }
+        if (count == infoMaterialVOS.size() && a > 0) {
+            return Result.success(advInfoUpdateVO);
+        } else {
+            throw new AdvInfoException(ExceptionConstants.INFO_EXCEPTION_INSERT_FAIL);
+        }
     }
 }
